testing: move assert_print_op definition to test_printer
